+++ AmazonChimeSDKDemo/AmazonChimeSDKDemo/MeetingModel.swift
@@ -41,9 +41,24 @@ class MeetingModel: NSObject {
                                                          cameraCaptureSource: videoModel.customSource)
     let uuid = UUID()
     var call: Call?
-
+    var isAppInBackground: Bool = false {
+        didSet {
+            if isAppInBackground {
+                wasLocalVideoOn = isLocalVideoActive
+                if wasLocalVideoOn {
+                    isLocalVideoActive = false
+                }
+                videoModel.pauseAllRemoteVideos()
+            } else {
+                if wasLocalVideoOn {
+                    isLocalVideoActive = true
+                }
+                videoModel.resumeAllRemoteVideosInCurrentPageExceptUserPausedVideos()
+            }
+        }
+    }
     private var savedModeBeforeOnHold: ActiveMode?
-
+    private var wasLocalVideoOn: Bool = false
     // States
     var activeMode: ActiveMode = .roster {
         didSet {